requirements.txt
Django==3.0.7
gunicorn==20.0.4
pylibmc==1.6.1
djangorestframework==3.11.0
requests==2.23.0
dj_database_url
django-debug-toolbar==2.2
sentry_sdk==0.14.4
Pillow==7.1.2
python-dotenv==0.11.0
